# docs/swagger.yaml
openapi: 3.0.0
paths:
  /devices:
    get:
      operationId: DeviceController_index
      summary: デバイスの一覧を取得
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceDetailDto'
      tags: &ref_0
        - device
    post:
      operationId: DeviceController_create
      summary: デバイスを登録
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpsertDto'
      responses:
        '201':
          description: ''
      tags: *ref_0
  /devices/{deviceId}:
    get:
      operationId: DeviceController_findOne
      summary: デバイスの詳細を取得
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDetailDto'
      tags: *ref_0
    put:
      operationId: DeviceController_update
      summary: デバイスを更新
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpsertDto'
      responses:
        '204':
          description: ''
      tags: *ref_0
    delete:
      operationId: DeviceController_remove
      summary: デバイスを削除
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
      responses:
        '204':
          description: ''
      tags: *ref_0
  /devices/{deviceId}/command/power:
    get:
      operationId: DeviceCommandController_getPower
      summary: 電源状態を取得
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceCommandPowerDetailDto'
      tags: &ref_1
        - command
    put:
      operationId: DeviceCommandController_setPower
      summary: 電源状態を変更
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCommandPowerRequestDto'
      responses:
        '204':
          description: ''
      tags: *ref_1
  /devices/{deviceId}/command/volume:
    get:
      operationId: DeviceCommandController_getVolume
      summary: 音量を取得
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceCommandVolumeDto'
      tags: *ref_1
    put:
      operationId: DeviceCommandController_setVolume
      summary: 音量を変更
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCommandVolumeDto'
      responses:
        '204':
          description: ''
      tags: *ref_1
  /devices/{deviceId}/command/mute:
    get:
      operationId: DeviceCommandController_getMute
      summary: ミュート状態を取得
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceCommandMuteDto'
      tags: *ref_1
    put:
      operationId: DeviceCommandController_setMute
      summary: ミュート状態を変更
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCommandMuteDto'
      responses:
        '204':
          description: ''
      tags: *ref_1
  /devices/{deviceId}/command/apps:
    get:
      operationId: DeviceCommandController_getApps
      summary: アプリの一覧を取得
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceCommandAppDetailDto'
      tags: *ref_1
  /devices/{deviceId}/command/launchApp/{productId}:
    post:
      operationId: DeviceCommandController_setApp
      summary: アプリを起動
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
        - name: productId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_1
  /devices/{deviceId}/command/sendKey/{vieraKey}:
    post:
      operationId: DeviceCommandController_setKey
      summary: キーを送信
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
        - name: vieraKey
          required: true
          in: path
          description: Viera Key
          schema:
            type: string
            enum:
              - NRC_POWER-ONOFF
              - NRC_RECLIST-ONOFF
              - NRC_DISP_MODE-ONOFF
              - NRC_CHG_INPUT-ONOFF
              - NRC_DATA-ONOFF
              - NRC_APPS-ONOFF
              - NRC_MENU-ONOFF
              - NRC_BLUE-ONOFF
              - NRC_RED-ONOFF
              - NRC_GREEN-ONOFF
              - NRC_YELLOW-ONOFF
              - NRC_SUBMENU-ONOFF
              - NRC_HOME-ONOFF
              - NRC_EPG-ONOFF
              - NRC_RETURN-ONOFF
              - NRC_UP-ONOFF
              - NRC_DOWN-ONOFF
              - NRC_LEFT-ONOFF
              - NRC_RIGHT-ONOFF
              - NRC_ENTER-ONOFF
              - NRC_NET_TD-ONOFF
              - NRC_NET_BS-ONOFF
              - NRC_NET_CS-ONOFF
              - NRC_D1-ONOFF
              - NRC_D2-ONOFF
              - NRC_D3-ONOFF
              - NRC_D4-ONOFF
              - NRC_D5-ONOFF
              - NRC_D6-ONOFF
              - NRC_D7-ONOFF
              - NRC_D8-ONOFF
              - NRC_D9-ONOFF
              - NRC_D0-ONOFF
              - NRC_D11-ONOFF
              - NRC_D12-ONOFF
              - NRC_CANCEL-ONOFF
              - NRC_MUTE-ONOFF
              - NRC_CH_UP-ONOFF
              - NRC_CH_DOWN-ONOFF
              - NRC_VOLUP-ONOFF
              - NRC_VOLDOWN-ONOFF
              - NRC_SKIP_PREV-ONOFF
              - NRC_SKIP_NEXT-ONOFF
              - NRC_REW-ONOFF
              - NRC_FF-ONOFF
              - NRC_REC-ONOFF
              - NRC_STOP-ONOFF
              - NRC_PAUSE-ONOFF
              - NRC_PLAY-ONOFF
              - NRC_MPX-ONOFF
              - NRC_ECO-ONOFF
              - NRC_VIERA_LINK-ONOFF
              - NRC_30S_SKIP-ONOFF
              - NRC_GUIDE-ONOFF
              - NRC_SPLIT-ONOFF
              - NRC_CC-ONOFF
              - NRC_OFFTIMER-ONOFF
              - MACRO_TV-ONOFF
              - MACRO_VIDEO-ONOFF
              - MACRO_HDMI2-ONOFF
              - MACRO_HDMI3-ONOFF
              - MACRO_HDMI4-ONOFF
      responses:
        '200':
          description: ''
      tags: *ref_1
  /devices/{deviceId}/auth/displayPinCode:
    post:
      operationId: DeviceAuthController_displayPinCode
      summary: PINコードをデバイスの画面に表示
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
      responses:
        '200':
          description: ''
        '503':
          description: デバイスの電源が入っていない場合
      tags: &ref_2
        - auth
  /devices/{deviceId}/auth:
    put:
      operationId: DeviceAuthController_auth
      summary: PINコード認証
      parameters:
        - name: deviceId
          required: true
          in: path
          description: Device ID
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceAuthDto'
      responses:
        '200':
          description: PINコードが正しい場合
        '400':
          description: PINコードが誤っている場合
      tags: *ref_2
info:
  title: VIERA Unofficial Web API
  description: VIERA Unofficial Web API
  version: '2.0'
  contact: {}
  license:
    name: ISC
    url: https://licenses.opensource.jp/ISC/ISC.html
tags: []
servers:
  - url: http://localhost:3003/v2
  - url: http://raspberrypi1.local:3003/v2
components:
  schemas:
    DeviceDetailDto:
      type: object
      properties:
        deviceId:
          type: number
        deviceName:
          type: string
        host:
          type: string
      required:
        - deviceId
        - deviceName
        - host
    DeviceUpsertDto:
      type: object
      properties:
        deviceName:
          type: string
        host:
          type: string
      required:
        - deviceName
        - host
    DeviceCommandPowerDetailDto:
      type: object
      properties:
        state:
          type: string
          enum:
            - 'ON'
            - 'OFF'
      required:
        - state
    DeviceCommandPowerRequestDto:
      type: object
      properties:
        state:
          type: string
          enum:
            - 'ON'
            - 'OFF'
            - TOGGLE
      required:
        - state
    DeviceCommandVolumeDto:
      type: object
      properties:
        value:
          type: number
      required:
        - value
    DeviceCommandMuteDto:
      type: object
      properties:
        value:
          type: boolean
      required:
        - value
    DeviceCommandAppDetailDto:
      type: object
      properties:
        label:
          type: string
        productId:
          type: string
        icon:
          type: string
      required:
        - label
        - productId
        - icon
    DeviceAuthDto:
      type: object
      properties:
        pinCode:
          type: string
      required:
        - pinCode